Docker 구동을 위한 redis 설정 추가
EmbededdRedisConfig 추가
RedisRepositoryConfig -> RedisConfig
build.gradle -> embedded-redis dependency 추가
RedisConfig -> 임시로 비밀번호 설정 막음
